@@ -3,11 +3,11 @@ description: Flutter plugin for reading and writing simple key-value pairs.
   Wraps NSUserDefaults on iOS and SharedPreferences on Android.
 repository: https://github.com/flutter/packages/tree/main/packages/shared_preferences/shared_preferences
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+shared_preferences%22
-version: 2.2.3
+version: 2.3.0
 
 environment:
-  sdk: ^3.2.0
-  flutter: ">=3.16.0"
+  sdk: ^3.3.0
+  flutter: ">=3.19.0"
 
 flutter:
   plugin:
@@ -28,12 +28,12 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  shared_preferences_android: ^2.1.0
-  shared_preferences_foundation: ^2.3.5
-  shared_preferences_linux: ^2.2.0
-  shared_preferences_platform_interface: ^2.3.0
-  shared_preferences_web: ^2.1.0
-  shared_preferences_windows: ^2.2.0
+  shared_preferences_android: ^2.3.0
+  shared_preferences_foundation: ^2.5.0
+  shared_preferences_linux: ^2.4.0
+  shared_preferences_platform_interface: ^2.4.0
+  shared_preferences_web: ^2.4.0
+  shared_preferences_windows: ^2.4.0
 
 dev_dependencies:
   flutter_test:
